ranks_* files: long -> int32 (half-size rank I/O; smaller update arrays for better cache)

+++ src/algorithm.h
@@ -19,13 +19,13 @@ typedef struct rank_pos_pair {
 
 int generate_local_runs (char * rank_dir, char * runs_dir, int total_chunks, int chunk_id, int h,
                          long working_chunk_size,
-                         long * current_ranks_buffer, long * next_ranks_buffer,
+                         int32_t * current_ranks_buffer, int32_t * next_ranks_buffer,
                          int * sa_buffer, RunRecord * runs_buffer);
 
 int resolve_global_ranks (char *temp_dir );
 int update_local_ranks (char * rank_dir, char * temp_dir, int total_chunks, int chunk_id, int h,
                         long working_chunk_size,
-                        long * buffer_current, long * buffer_next,
+                        int32_t * buffer_current, int32_t * buffer_next,
                         int * sa_buffer, long * updated_ranks);
 
 #endif
